requirements.txt

pytest==8.3.2
httpx
pyyaml


requests==2.32.3


pillow==10.4.0

semantic-kernel==0.3.13.dev0
openai==0.28.1
fastapi==0.110.0
uvicorn==0.29.0
pydantic==1.10.13
python-dotenv==1.0.0
azure-identity==1.15.0
